Adding ImageLoader Component
Added a loading and error msg states to the app, referring to the AJAX request loading the data, While the data is loading you should see a ‘loading…’ message, If the data fetching failed an error message on the screen will be shown.
Added a new ImageLoader component with its own internal state and re-fetch mechanism.
The ImageLoader component is used in the Card component instead of the <Thumb> component. The ImageLoader component has its own <img> element which has a default icon of a robot as long as the image is loading. Once the image is loaded successfully it replaces the icon. Using onLoad and onError on the IMG JSX element to implement this functionality.